File loading
Instead of saving each frame every iteration, instead, at the start of the program, each frame is taken out and saved as a single file. The bottleneck is still talking with Windows and updating which image to use. Seems there is a limit to how fast the background image can change.